test(router-core): move search context boundary coverage to core
